Refactoring audio code
- Add AudioBus
- Move resampling to AudioBus
- Move SincResampler to engine/audio

+++ src/engine/audio/audio_sink.h
@@ -15,7 +15,7 @@ class AudioSink {
   virtual void Suspend() = 0;
   virtual void Resume() = 0;
 
-  virtual int GetHardwareSampleRate() = 0;
+  virtual size_t GetHardwareSampleRate() = 0;
 
  private:
   AudioSink(const AudioSink&) = delete;
